added sub dir for types and canvas
