+++ src/types/input_types.rs
@@ -1,0 +1,14 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum MouseButton {
+    Left,
+    Right,
+    Middle,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum ScrollAxis {
+    Up,
+    Down,
+    Left,
+    Right,
+}
